refactor: winner 테스트 생성자로 변경
