+++ src/test/java/racewinner/domain/CarsTest.java
@@ -13,16 +13,16 @@ public class CarsTest {
     @Test
     void winnerTest() {
         //given
-        final Car bobo = new Car("bobo", () -> true);
-        final Car coco = new Car("coco", () -> true);
-        final Car nono = new Car("nono", () -> false);
+        final Car bobo = new Car("bobo", new Position(3),  null);
+        final Car coco = new Car("coco", new Position(3), null);
+        final Car nono = new Car("nono", new Position(1), null);
         final Cars cars = new Cars(Arrays.asList(bobo, coco, nono));
 
         //when
-        cars.move();
         List<String> winnerList = cars.findWinner();
 
         //then
+        assertThat(winnerList.size()).isEqualTo(2);
         assertThat(winnerList).contains("bobo");
         assertThat(winnerList).contains("coco");
     }
