+++ src/test/java/racewinner/factory/CarsFactoryTest.java
@@ -2,7 +2,7 @@ package racewinner.factory;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racewinner.domain.factory.CarsFactory;
+import racewinner.domain.CarsFactory;
 import racewinner.exception.InputEmptyException;
 import racewinner.exception.InvalidDelimiterException;
